いいね非同期未完了

+++ app/controllers/favorites_controller.rb
@@ -8,8 +8,10 @@ class FavoritesController < ApplicationController
 
   def destroy
   book = Book.find(params[:book_id])
-    favorite = current_user.favorites.new(book_id: book.id)
+  favorite = current_user.favorites.find_by(book_id: book.id)
+  if favorite.present?
     favorite.destroy
-    redirect_to request.referer
+  end
+  redirect_to request.referer
   end
 end
